requirements.txt
# Core LangChain
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66

# LLM & Vector Store Integrations
langchain-groq==0.2.1
langchain-pinecone==0.2.0
langchain-huggingface==0.1.2

# Pinecone Vector Database
pinecone-client==5.0.1

# Embeddings
sentence-transformers==3.3.1

# Document Processing
pypdf==5.1.0

# Web Framework
flask==3.1.1

# Environment Variables
python-dotenv==1.1.0

# Additional Dependencies
tiktoken==0.8.0
huggingface-hub==0.26.2


streamlit

# Local package
-e .
